test(codec): relocate plan wire tests and adjust the core callers
Move the plan serialization/deserialization wire tests from core into the codec
module (where the conversions now live), split per origin, and rewrite the two
core tests that exercised Plan.ToProto directly so core no longer references the
moved methods. Domain-only tests stay in core.

Part of #280.
