+++ codec/testdata/lambda/basic-lambda.json
