+++ codec/testdata/lambda/lambda-with-function.json
