+++ codec/testdata/lambda/lambda-with-list-transform.json
