+++ codec/testdata/lambda/nested-lambda-outer-ref.json
