+++ codec/testdata/plan/ctas_basic.json
