+++ codec/testdata/plan/ctas_with_filter.json
