+++ codec/testdata/plan/insert_from_select.json
