+++ codec/testdata/plan/value_with_literal.json
