+++ codec/testdata/plan/value_with_scalar.json
