+++ plan/virtual_table_from_expr_test.go
@@ -1,13 +1,11 @@
 package plan_test
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/substrait-io/substrait-go/v9/expr"
-	"github.com/substrait-io/substrait-go/v9/extensions"
 	"github.com/substrait-io/substrait-go/v9/plan"
 	"github.com/substrait-io/substrait-go/v9/types"
 )
@@ -17,27 +15,8 @@ var (
 	v2 = expr.PrimitiveLiteral[int32]{Value: 2, Type: &types.Int32Type{Nullability: types.NullabilityRequired}}
 )
 
-// makeAddExpr constructs expression val1 + val2.
-func makeAddExpr(t *testing.T, b plan.Builder, val1, val2 expr.Literal) expr.Expression {
-	id := extensions.FunctionID{
-		URN:  "extension:io.substrait:functions_arithmetic",
-		Name: "add:i32_i32",
-	}
-	b.GetFunctionRef(id.URN, id.Name)
-	scalarExpr, err := b.ScalarFn(id.URN, id.Name, nil, val1, val2)
-	require.NoError(t, err)
-	return scalarExpr
-}
-
 func buildLiteralExpressions(_ *testing.T, _ plan.Builder) []expr.VirtualTableExpressionValue {
 	return []expr.VirtualTableExpressionValue{{&v1, &v2}}
-}
-
-// buildScalarAddExpression builds a scalar binary add expression
-func buildScalarAddExpression(t *testing.T, b plan.Builder) []expr.VirtualTableExpressionValue {
-	s1 := makeAddExpr(t, b, &v1, &v1)
-	s2 := makeAddExpr(t, b, &v2, &v2)
-	return []expr.VirtualTableExpressionValue{{s1, s2}}
 }
 
 func TestVirtualTableFromExpr(t *testing.T) {
@@ -54,33 +33,4 @@ func TestVirtualTableFromExpr(t *testing.T) {
 	vtRemap, err := vt.Remap(1)
 	require.NoError(t, err)
 	assert.Equal(t, "struct<i32>", vtRemap.RecordType().String())
-}
-
-// TestNamedTableInsertRoundTrip verifies that generated plans match the expected JSON.
-func TestVirtualTableFromExprRoundTrip(t *testing.T) {
-	for _, td := range []struct {
-		name             string
-		fieldNames       []string
-		buildExprForTest func(t *testing.T, b plan.Builder) []expr.VirtualTableExpressionValue
-	}{
-		{"value_with_literal", []string{"col0", "col1"}, buildLiteralExpressions},
-		{"value_with_scalar", []string{"col0", "col1"}, buildScalarAddExpression},
-	} {
-		t.Run(td.name, func(t *testing.T) {
-			// Load the expected JSON. This will be our baseline for comparison.
-			expectedJson, err := testdata.ReadFile(fmt.Sprintf("testdata/%s.json", td.name))
-			require.NoError(t, err)
-
-			// build plan for Project with virtual table
-			b := plan.NewBuilderDefault()
-			valueExpr := td.buildExprForTest(t, b)
-			virtualTableExpr, err := b.VirtualTableFromExpr(td.fieldNames, valueExpr...)
-			require.NoError(t, err)
-			virtualTablePlan, err := b.Plan(virtualTableExpr, td.fieldNames)
-			require.NoError(t, err)
-
-			// Check that the generated plan matches the expected JSON.
-			checkRoundTrip(t, string(expectedJson), virtualTablePlan)
-		})
-	}
 }
